Migrate project into web app
The app has an address autocompletion input field that accepts any address in Aarhus kommune and fetches the pickup dates from a local SQLite DB or directly from MitAffald as a fallback.

The backend is written in Python as a Flask/SQLAlchemy app (based on the original script). The frontend is built in React with `create-react-app`.

Requests to MitAffald cannot be sent directly from the frontend to MitAffald due to CORS, but that is also unnecessary as we want the backend to check if it has the content persisted already. The backend is not restricted by CORS so can request the data without any issues, but a DB with all addresses in Aarhus has been imported and pre-ingested. This both serves to significantly speed up the initial view, but also ensures that we have data for (almost) the entire year and not only the future (which is the only thing that MitAffald exposes).

A docker deployment containing both the front- and backend (which serves the frontend) is included.

Google Calendar integration will be added later at which point the original script (which is moved into `tools`) may get removed.