Fix Chrome date parsing "bug"
Chrome browsers apparently parse dates where the month or day is single-digit in different time zones dependening on whether the values are padded with a leading zero or not. If it is, the value is parsed in UTC, if it isn't, it's parsed in the local time zone. I agree, that's pretty fucked...

With this commit, we're parsing the values into numbers manually and parsing them to `Date.UTC` instead. That's the same function as used in 'buildCalendarData', so that should (as if you can ever know with JS) ensure that you get the same result in all engines.