Fix date parsing of valid time
The 'validTime' field of the calendar data uses the same format as the calendar dates but didn't get the patch in 40170c. Had the error just been an "off by timezone" again it wouldn't have been a big deal since would at most render a single day (around the time of 'validTime') in the wrong "valid" state.

But it turns out that JavaScript's date parsing is even more broken than previously observed: On Safari (iOS at least), 'Date.parse' returns NaN if leading zeros aren't included. Therefore *all* days were renderes as "invalid"..!

The fix is to extract the patch of the calendar data into a function and reuse it for the 'validTime' data.