Separate import from query (#2)
Don't query MitAffald when data is missing. Instead, add a script 'import_calendars.py' and return empty data or error when the lookup fails. Also, the script records NULL values in the DB for unsuccessful imports as that indicates that the data was attempted imported but failed.

The frontend was updated to handle calendar dates being null.

The backend also updated to no longer let 404 responses map to 'index.html'.