Frontend: Resolve and use current year
The app just assumed the year to be 2023 and crashed if it wasn't ("today" resolved to 'undefined' which would trigger a "null" pointer exception later on).

This change required adding logic for computing the necessary properties for rendering the calendar (whether it's a leap year and the weekday of Jan 1st) as the info was just hardcoded as constants for 2023.

The backend doesn't have any data for 2024 because MitAffald is down for God knows how long now. This also means that these changes haven't been truly tested yet.

Library dependencies were bumped as well.